refactor(IAudioOutput): consolidate playback position estimation into IAudioOutput
IAudioOutput becomes a base class instead of a pure interface:
backends (Esp32I2sAudioOutput, NativeI2sAudioOutput) now implement
private doBegin/doWrite/doStop, while the public begin/write/stop
wrappers in the base own all play-position estimation logic that
was previously duplicated between Esp32A2dpBluetoothSource and
NativeA2dpBluetoothSource.

- Add PLAY_STATUS enum and _updatePlayingState() /
  _resetPlayPositionEstimate() / _updateLastRealPositionTimestamp() /
  _getEstimatedPosition() on IAudioOutput, so both BT backends drive
  the estimator through one interface instead of keeping their own
  _isPlaying / _rawAudioDataBytesReceived / _lastRealPositionTimestamp
  members.
- Esp32A2dpBluetoothSource / NativeA2dpBluetoothSource: drop the
  now-dead per-backend copies of that state; next()/previous()/
  metadata-title handling and _avrcPlayPosTrampoline now just forward
  to _audioOutput->_resetPlayPositionEstimate() /
  _updateLastRealPositionTimestamp(). _readDataStreamTrampoline /
  _streamReaderTrampoline call _audioOutput->write() and (still under
  TRACK_POSITION_FIX) pull the estimate back via
  _getEstimatedPosition() to feed onPlayPosition().
- Esp32A2dpBluetoothSource::begin(): set_stream_reader(...) is now
  registered unconditionally instead of only under
  #ifdef TRACK_POSITION_FIX, since this is the actual PCM routing
  path into IAudioOutput and must not depend on that macro.
- Esp32I2sAudioOutput / NativeI2sAudioOutput: rename begin/write/stop
  overrides to doBegin/doWrite/doStop to match the new base class
  contract.
- storage::setBool(): cast to uint8_t before NVS.setInt() to avoid
  the implicit bool->int widening.

+++ src/storage.cpp
@@ -19,7 +19,7 @@ namespace storage
 
     void setBool(const char *key, bool value)
     {
-        if (!NVS.setInt(key, value ? 1 : 0, true))
+        if (!NVS.setInt(key, (uint8_t) (value ? 1 : 0), true))
         {
             ESP_LOGE("storage", "failed to save %s to NVS", key);
         }
